deployment: always restart service on reconfigure
The previous code tried to determine when a restart was necessary to
provide online reconfiguration. Unfortunately, this code wasn't kept
up-to-date, and is less necessary now that we have the `debug` server
where we can put functionality such as `set-log-level`.

We currently believe that every configuration item in the "System"
configuration that is actually considered in the converger will render
a configuration file, so the self-hup code was removed.

We return to idle after 300 seconds just in case hab is in trouble.

Signed-off-by: Steven Danna <steve@chef.io>