Fix IAM V2 bug for new A1 data collector token entries (#2961)
Since this commit https://github.com/chef/automate/commit/a2e1a87bc749cd89d5011c82b5379d6fdf8fdc7f we have been enforcing project permissions validation on token entries. When the authn server starts up, if it's configured with a legacy A1 data collector token, it will first check if it exists, then insert a token with the config specified value if it doesn't. However, on IAM V2, it performs validation that the subject from context has permissions on the projects specified. In this case, we do not wish to validate projects because 1) we are just inserting the token in the default project (will always exist) and 2) this is a domain specific operation being initiated by the server which shouldn't need to round-trip to authz to ask for permission.

Signed-off-by: Tyler Cloke <tylercloke@gmail.com>